--- log/printf.go
@@ -1,21 +1,0 @@
-// +build !log_linenumber
-
-package log
-
-import "os"
-
-func (logger *Logger) doPrintf(level int, printLevel string, format string, a ...interface{}) {
-	if level < logger.level {
-		return
-	}
-	if logger.baseLogger == nil {
-		panic("logger closed")
-	}
-
-	format = printLevel + format
-	logger.baseLogger.Printf(format, a...)
-
-	if level == fatalLevel {
-		os.Exit(1)
-	}
-}
