--- log/printf_linenumber.go
@@ -1,35 +1,0 @@
-// +build log_linenumber
-
-package log
-
-import (
-	"os"
-	"runtime"
-	"strconv"
-)
-
-func logSite() string {
-	_, file, line, ok := runtime.Caller(4)
-	if !ok {
-		file = "???"
-		line = 0
-	}
-	c := string(file + ":" + strconv.FormatInt(int64(line), 10))
-	return c
-}
-
-func (logger *Logger) doPrintf(level int, printLevel string, format string, a ...interface{}) {
-	if level < logger.level {
-		return
-	}
-	if logger.baseLogger == nil {
-		panic("logger closed")
-	}
-
-	format = printLevel + "["+logSite()+"] " + format
-	logger.baseLogger.Printf(format, a...)
-
-	if level == fatalLevel {
-		os.Exit(1)
-	}
-}
